Move all console state logic into Console.js
