+++ src/components/tabs/console/Console.js
@@ -1,85 +1,34 @@
-import { h } from 'preact';
+import { createRef, h } from 'preact';
+import { useEffect, useReducer } from 'preact/hooks';
 import styles from './Console.module.scss';
 import Prompt from './Prompt';
 import CopyButton from '../../CopyButton';
 import consoleHook, { console } from '../../../utilities/hooks/consoleHook';
-import { useCallback, useState } from 'preact/hooks';
 import getTimestamp from '../../../utilities/getTimestamp';
 import Tree, { getNode } from '../../dataTree/Tree';
 
 /**
- * The content of the Console tab
+ * The content and logic for the Console tab
  */
 export default () => {
 
   // Store the data for every console call
-  const [rows, setRows] = useState([]);
+  const [rows, addRow] = useReducer((rows, row) => rows.concat(row), []);
+  consoleHook.onConsole((...args) => addRow(parseConsoleArgs(...args)));
 
-  // Send all console output calls here
-  consoleHook.onConsole((...args) => addRow(...args));
-  // consoleHook.onConsole(addRow); // TODO: Use this?
-
-  /**
-   * Add the data from the console call to our rows state var
-   */
-  const addRow = useCallback(({ level = 'log', args, stack = [] }) => {
-    const timestamp = getTimestamp();
-
-    const frame = (stack && stack[0]) || {};
-    const fileName = frame.fileName && frame.lineNumber
-      ? `${frame.fileName}:${frame.lineNumber}`
-      : frame.fileName || '';
-
-    // TODO: Make this part better (treeData vs <Tree>)
-    const treeData = [];
-    const argElements = Object.keys(args).map((key) => {
-      const arg = args[key];
-      const isError = arg instanceof Error ||
-        (arg && arg.constructor && arg.constructor.name && arg.constructor.name.indexOf('Error') > -1);
-      const dataTree = getNode(isError ? (arg.stack || arg) : arg);
-      treeData.push(dataTree);
-      return <Tree dataTree={dataTree}/>;
-    });
-
-    // Append do the DOM
-    const el = {};
-    if (el.scrollIntoView) {
-      el.scrollIntoView();
-    }
-
-    console.log('Current rows:', rows);
-    setRows([...rows, {
-      argElements,
-      fileName,
-      level,
-      stack,
-      stackString: stack.map(frame => frame.path).join('\n'),
-      timestamp,
-      treeData,
-    }]);
-  }, [rows]);
-
-  const toJson = () => JSON.stringify({
-    userAgent: navigator.userAgent,
-    href: window.location.href,
-    time: getTimestamp(),
-    consoleOutput: outputData.map(rowData => {
-      const treeData = rowData.treeData.map(tree => JSON.parse(tree.toJson()));
-      return {
-        time: rowData.timestamp,
-        stack: rowData.stack,
-        output: treeData
-      };
-    })
-  });
+  // Scroll to the bottom on render
+  const outputRef = createRef();
+  useEffect(() => outputRef.current.scrollTop = outputRef.current.scrollHeight);
 
   return (
     <div className={styles.console}>
-      <div className={styles.output}>
+      <div className={styles.output} ref={outputRef}>
         {rows.map(row => (
           <div className={`${styles.row} ${styles[row.level]}`}>
             <div className={styles.timestamp}>{row.timestamp.split(' ')[1]}</div>
-            <div className={styles.consoleArgs}>{[...row.argElements]}</div>
+            <div className={styles.consoleArgs}>{
+              row.argData.map(dataTree => <Tree dataTree={dataTree}/>)
+            }</div>
             <div className={styles.fileName}>
               <CopyButton getText={() => row.stackString} title='Copy stack'/>
               <span title={row.stackString}>{row.fileName}</span>
@@ -88,9 +37,117 @@ export default () => {
         ))}
       </div>
       <div className={styles.copyAllButton}>
-        <CopyButton getText={() => test()} title="Copy all output"/>
+        <CopyButton getText={() => toJson(rows)} title="Copy all output"/>
       </div>
-      <Prompt/>
+      <Prompt exec={exec} getHistory={getHistory}/>
     </div>
   );
 };
+
+/**
+ * Convert consoleHook data to row state data
+ */
+export const parseConsoleArgs = ({ level = 'log', args, stack = [] }) => {
+  const timestamp = getTimestamp();
+
+  const frame = (stack && stack[0]) || {};
+  const fileName = frame.fileName && frame.lineNumber
+    ? `${frame.fileName}:${frame.lineNumber}`
+    : frame.fileName || '';
+
+  const argData = Object.keys(args).map((key) => {
+    const arg = args[key];
+    const isError = arg instanceof Error ||
+      (arg && arg.constructor && arg.constructor.name && arg.constructor.name.indexOf('Error') > -1);
+    return getNode(isError ? (arg.stack || arg) : arg);
+  });
+
+  return {
+    argData,
+    fileName,
+    level,
+    stack,
+    stackString: stack.map(frame => frame.path).join('\n'),
+    timestamp,
+  };
+};
+
+/**
+ * Generate a json representation of everything in the component
+ */
+export const toJson = (rows) => JSON.stringify({
+  userAgent: navigator.userAgent,
+  href: window.location.href,
+  time: getTimestamp(),
+  consoleOutput: rows.map(row => {
+    const argData = row.argData.map(dataTree => JSON.parse(dataTree.toJson()));
+    return {
+      time: row.timestamp,
+      stack: row.stack,
+      output: argData
+    };
+  })
+});
+
+/**
+ * Local storage key for command history
+ */
+const historyKey = 'fTwelve.history';
+
+/**
+ * Push a command onto history and write to local storage
+ */
+export const setHistory = (command, maxSize = 50) => {
+  history.unshift(command);
+  history.splice(maxSize);
+  if (window.localStorage) {
+    window.localStorage.setItem(historyKey, JSON.stringify(history));
+  }
+};
+
+/**
+ * Retrieve all history items
+ */
+export const getHistory = () => window.localStorage
+  ? (JSON.parse(window.localStorage.getItem(historyKey)) || [])
+  : history || [];
+
+/**
+ * Push a console command onto the history array and execute it
+ */
+export const exec = (command) => {
+  setHistory(command);
+  window.console.log(command);
+  try {
+    window.console.log(parseCommand(command));
+  } catch (e) {
+    window.console.error(e);
+  }
+};
+
+/**
+ * Parse a string and safely evaluate it in JS (as opposed to `eval` or `Function`)
+ */
+export const parseCommand = (command) => {
+  command = command.trim();
+  if (command.match(/^".*"$/) || command.match(/^'.*'$/)) {
+    return command.slice(1, -1);
+  }
+  const expressions = command.split(/\s*=\s*/);
+  const firstExpression = expressions.shift();
+  return firstExpression.replace(/(?=\[)/g, '.').split('.').reduce((object, memberString, idx, array) => {
+    const bracketMatch = memberString.match(/^\[([^\]]*)]$/);
+    const memberName = bracketMatch ? bracketMatch[1].replace(/^["']|["']$/g, '') : memberString;
+    if (expressions.length > 0 && idx === array.length - 1) {
+      // If there are things to the right of the equals sign, assign it to the left
+      (object || {})[memberName] = parseCommand(expressions.join('='));
+    }
+    return (object || {})[memberName];
+  }, window);
+};
+
+/**
+ * Array of recently executed commands
+ * Note: This is not a state variable
+ */
+const history = getHistory();
